pyproject.toml
[project]
name = "cloudgripper-api"
requires-python = ">=3.11"
version = "0.1.0"
dependencies = [
    "flask>=3.1.0",
    "imageio==2.22.1",
    "imageio-ffmpeg==0.4.7",
    "numpy==1.26.4",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pynput>=1.8.1",
    "pytest==7.3.1",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "scipy>=1.13.1",
    "tqdm>=4.67.1",
]

[tool.uv]
environments = [
    "sys_platform == 'linux'",
]

